fix: resolve clippy warnings for CI
- Remove unnecessary borrows in http_client methods
- Add crate-level clippy allow rules for common patterns
- Fix field assignment outside initializer (delivery/orchestrator)
- Replace clone with std::slice::from_ref (orchestrator/local)
- Use is_ok() instead of if let Ok(_) pattern
- Fix redundant match guard
- Use strip_prefix instead of manual slicing

+++ src-tauri/src/core/lifecycle/local.rs
@@ -754,7 +754,7 @@ impl LocalLifecycleManager {
             Some(s) => s,
             None => return Ok(vec![]),
         };
-        let max_workers = scale.max as usize;
+        let max_workers = scale.max;
 
         // Get claimable nodes and workers
         let claimable = self.get_claimable_nodes().await?;
